Fixes apply button logic in wallet, abstracts SuccessMessage in applications
